Adds React Redux and Updates Dependencies
The dependencies have been updated to the latest versions. React Redux has been added to manage the blogs and consoles; as well as to remove props drilling efforts. More logic has been added to make blogs on the homepage generate random featured articles and posts in the More Posts section.